ndm: make sure `github.com/invopop/jsonschema` is not in the final binary
