+++ pkg/networkdevice/profile/profiledefinition/metadata.go
@@ -5,18 +5,11 @@
 
 package profiledefinition
 
-import "github.com/invopop/jsonschema"
-
 // MetadataDeviceResource is the device resource name
 const MetadataDeviceResource = "device"
 
 // MetadataConfig holds configs per resource type
 type MetadataConfig ListMap[MetadataResourceConfig]
-
-// JSONSchema defines the JSON schema for MetadataConfig
-func (mc MetadataConfig) JSONSchema() *jsonschema.Schema {
-	return ListMap[MetadataResourceConfig](mc).JSONSchema()
-}
 
 // MarshalJSON marshals the metadata config
 func (mc MetadataConfig) MarshalJSON() ([]byte, error) {
